feat(app): split predictions into new hooks for submision and retrival and update usage
