+++ client/hooks/useHomeFixtures.ts
@@ -2,16 +2,10 @@ import { useQuery } from '@tanstack/react-query'
 import { API_URL } from '../lib/constants'
 import axios from 'axios'
 
-interface FixtureDateRange {
-  startDate: string
-  endDate: string
-}
-
-const getFixtures = async ({ startDate, endDate }: FixtureDateRange) => {
+// Returns all fixtures on the hompage (live, upcoming, finished) will eventually be merged with useFixtures pending API
+export const getHomepageFixtures = async () => {
   try {
-    const { data } = await axios.get(
-      `${API_URL}/fixtures-all/date?start=${startDate}&end=${endDate}`
-    )
+    const { data } = await axios.get(`${API_URL}/fixtures`)
     return data
   } catch (error: unknown) {
     if (error instanceof Error) {
@@ -24,6 +18,8 @@ const getFixtures = async ({ startDate, endDate }: FixtureDateRange) => {
   }
 }
 
-export const useFixtures = ({ startDate, endDate }: FixtureDateRange) => {
-  return useQuery(['fixtures'], () => getFixtures({ startDate, endDate }))
+export const useHomepageFixtures = () => {
+  return useQuery(['homepageFixtures'], getHomepageFixtures, {
+    refetchInterval: 30000,
+  })
 }
